refactor: DTO 패키징 분리
